Update 9 Publication.cpp
In this program, while catching the error, if we didn't specified what we have to check then it may cause ambiguity in the program sometimes. and can lead to form error. So it is very much necessary to specify particular value of variable to catch and perform operation in after catching an error.